Forward compatibility for MySQL-like platforms
