+++ src/Platforms/Keywords/MariaDBKeywords.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Doctrine\DBAL\Platforms\Keywords;
+
+abstract class MariaDBKeywords extends MySQLKeywords
+{
+}
