+++ src/Platforms/Keywords/MariaDb102Keywords.php
@@ -7,7 +7,7 @@ namespace Doctrine\DBAL\Platforms\Keywords;
  *
  * @link https://mariadb.com/kb/en/the-mariadb-library/reserved-words/
  */
-final class MariaDb102Keywords extends MySQLKeywords
+final class MariaDb102Keywords extends MariaDBKeywords
 {
     public function getName(): string
     {
