+++ src/Platforms/MariaDBPlatform.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Doctrine\DBAL\Platforms;
+
+/**
+ * Provides the behavior, features and SQL dialect of the MariaDB database platform of the oldest supported version.
+ */
+abstract class MariaDBPlatform extends MySQLPlatform
+{
+}
