+++ src/Schema/MySQLSchemaManager.php
@@ -2,9 +2,9 @@
 
 namespace Doctrine\DBAL\Schema;
 
+use Doctrine\DBAL\Platforms\AbstractMySQLPlatform;
 use Doctrine\DBAL\Platforms\MariaDb1027Platform;
 use Doctrine\DBAL\Platforms\MySQL;
-use Doctrine\DBAL\Platforms\MySQLPlatform;
 use Doctrine\DBAL\Types\Type;
 
 use function array_change_key_case;
@@ -23,7 +23,7 @@ use const CASE_LOWER;
 /**
  * Schema manager for the MySQL RDBMS.
  *
- * @extends AbstractSchemaManager<MySQLPlatform>
+ * @extends AbstractSchemaManager<AbstractMySQLPlatform>
  */
 class MySQLSchemaManager extends AbstractSchemaManager
 {
@@ -168,27 +168,27 @@ class MySQLSchemaManager extends AbstractSchemaManager
                 break;
 
             case 'tinytext':
-                $length = MySQLPlatform::LENGTH_LIMIT_TINYTEXT;
+                $length = AbstractMySQLPlatform::LENGTH_LIMIT_TINYTEXT;
                 break;
 
             case 'text':
-                $length = MySQLPlatform::LENGTH_LIMIT_TEXT;
+                $length = AbstractMySQLPlatform::LENGTH_LIMIT_TEXT;
                 break;
 
             case 'mediumtext':
-                $length = MySQLPlatform::LENGTH_LIMIT_MEDIUMTEXT;
+                $length = AbstractMySQLPlatform::LENGTH_LIMIT_MEDIUMTEXT;
                 break;
 
             case 'tinyblob':
-                $length = MySQLPlatform::LENGTH_LIMIT_TINYBLOB;
+                $length = AbstractMySQLPlatform::LENGTH_LIMIT_TINYBLOB;
                 break;
 
             case 'blob':
-                $length = MySQLPlatform::LENGTH_LIMIT_BLOB;
+                $length = AbstractMySQLPlatform::LENGTH_LIMIT_BLOB;
                 break;
 
             case 'mediumblob':
-                $length = MySQLPlatform::LENGTH_LIMIT_MEDIUMBLOB;
+                $length = AbstractMySQLPlatform::LENGTH_LIMIT_MEDIUMBLOB;
                 break;
 
             case 'tinyint':
