+++ tests/Schema/Visitor/SchemaSqlCollectorTest.php
@@ -2,7 +2,7 @@
 
 namespace Doctrine\DBAL\Tests\Schema\Visitor;
 
-use Doctrine\DBAL\Platforms\MySQLPlatform;
+use Doctrine\DBAL\Platforms\AbstractMySQLPlatform;
 use Doctrine\DBAL\Schema\Schema;
 use PHPUnit\Framework\TestCase;
 
@@ -10,9 +10,9 @@ class SchemaSqlCollectorTest extends TestCase
 {
     public function testCreateSchema(): void
     {
-        $platformMock = $this->getMockBuilder(MySQLPlatform::class)
+        $platformMock = $this->getMockBuilder(AbstractMySQLPlatform::class)
             ->onlyMethods(['getCreateTableSql', 'getCreateSequenceSql', 'getCreateForeignKeySql'])
-            ->getMock();
+            ->getMockForAbstractClass();
         $platformMock->expects(self::exactly(2))
                      ->method('getCreateTableSql')
                      ->willReturn(['foo']);
@@ -32,9 +32,9 @@ class SchemaSqlCollectorTest extends TestCase
 
     public function testDropSchema(): void
     {
-        $platformMock = $this->getMockBuilder(MySQLPlatform::class)
+        $platformMock = $this->getMockBuilder(AbstractMySQLPlatform::class)
             ->onlyMethods(['getDropTableSql', 'getDropSequenceSql', 'getDropForeignKeySql'])
-            ->getMock();
+            ->getMockForAbstractClass();
         $platformMock->expects(self::exactly(2))
                      ->method('getDropTableSql')
                      ->willReturn('tbl');
